requirements.txt
Django>=5.0,<6.0
django-htmx
psycopg2-binary
Pillow
faker
python-dateutil